Add files via upload
the scripts contain piplines to find Differential Abundance of Genera using DESeq2, radEmu, ALDEx2 (in RMD file), and ANCOMBC by qiime2 in bash script.
The compressed file contains plots for intersected and unique taxa between DESeq2, radEmu, ALDEx2. the number of taxa taken for plotting in species is 50 (except for some samples, the number is written in the txt file). The number of taxa taken for plotting genus is 25.